Moved CommandMessage to Cedar.CommandHandling
