+++ src/Cedar.CommandHandling/Properties/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyTitle("Cedar.CommandHandling")]
