+++ src/Cedar.HttpCommandHandling.Example/05_CommandVersioning.cs
@@ -12,6 +12,7 @@ namespace Cedar.HttpCommandHandling.Example.CommandVersioning
 {
     using System;
     using System.Threading.Tasks;
+    using Cedar.CommandHandling;
     using Cedar.HttpCommandHandling;
 
     // 1. A number of versions of the same command.
