+++ src/Cedar.HttpCommandHandling.Example/06_UnitTestingHandlers.cs
@@ -8,6 +8,7 @@ namespace Cedar.HttpCommandHandling.Example.UnitTestinHandlers
     using System;
     using System.Threading;
     using System.Threading.Tasks;
+    using Cedar.CommandHandling;
     using Cedar.HttpCommandHandling;
     using FakeItEasy;
     using Xunit;
