+++ src/Cedar.HttpCommandHandling.Example/Cedar.HttpCommandHandling.Example.csproj
@@ -80,6 +80,10 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\Cedar.CommandHandling\Cedar.CommandHandling.csproj">
+      <Project>{f6d3ad03-f93e-4b8b-b759-7a499ad4f4e4}</Project>
+      <Name>Cedar.CommandHandling</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Cedar.HttpCommandHandling.Client\Cedar.HttpCommandHandling.Client-PCL.csproj">
       <Project>{b3591387-2ebc-4d6f-a0c0-bea555dfd192}</Project>
       <Name>Cedar.HttpCommandHandling.Client-PCL</Name>
