+++ src/Cedar.HttpCommandHandling.Tests/CommandHandlingTests.cs
@@ -6,6 +6,7 @@
     using System.Net.Http;
     using System.Net.Http.Headers;
     using System.Threading.Tasks;
+    using Cedar.CommandHandling;
     using FluentAssertions;
     using Xunit;
     using Xunit.Extensions;
