+++ src/Cedar.HttpCommandHandling/CommandHandlerController.cs
@@ -9,6 +9,7 @@ namespace Cedar.HttpCommandHandling
     using System.Threading.Tasks;
     using System.Web.Http;
     using Cedar.Annotations;
+    using Cedar.CommandHandling;
     using Cedar.HttpCommandHandling.TypeResolution;
 
     internal class CommandHandlerController : ApiController
